@@ -11,53 +11,59 @@ description: A simple comic viewer application
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.4.9+37
+version: 1.4.10-beta.1+38
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  dio: "^3.0.0"
-  dio_proxy: ^1.0.10
+  dio: "^4.0.0"
   flutter_parallax: "^0.1.2"
-  cached_network_image: "^2.2.0"
-  flutter_cache_manager: ^1.4.1
-  dio_http_cache: '^0.2.0'
-  sqflite: ^1.3.0
-  flutter_easyloading: ^1.1.3
-  package_info: ^0.4.0+2
-  url_launcher: ^5.4.10
-  fluttericon: ^1.0.7
+  cached_network_image: "^3.0.0"
+  flutter_cache_manager: ^3.0.1
+  sqflite: ^2.0.0
+  dio_cache_interceptor: ^2.1.0
+  flutter_easyloading: ^3.0.0
+  package_info: ^2.0.0
+  url_launcher: ^6.0.3
+  fluttericon: ^2.0.0
   flutter_webview_plugin: ^0.3.11
-  markdown_widget: ^1.2.0
+  markdown_widget:
+    git: https://github.com/asjqkkkk/markdown_widget
   badges: ^1.1.6
   yin_drag_sacle: ^0.0.2
   flutter_swiper: ^1.1.6
-  event_bus: ^1.1.1
-  permission_handler: ^5.0.1
-  path_provider: ^1.6.16
-  provider: ^4.3.2+2
+  permission_handler: ^6.1.3
+  path_provider: ^2.0.0
+  provider: ^5.0.0
   toast: ^0.1.5
   uni_links: ^0.4.0
-  share: ^0.6.5+2
-  file_picker: ^2.1.5
+  share: ^2.0.1
+  file_picker: ^3.0.1
   flutter_downloader: ^1.5.2
   logger: ^0.9.4
-  flutter_widget_from_html: ^0.5.1+5
+  flutter_widget_from_html: ^0.6.0
   flutter_easyrefresh: ^2.1.8
-  connectivity: ^2.0.2
-  battery: ^1.0.7
-  flutter_spinkit: "^4.1.2"
-  version: ^1.2.0
+  connectivity: ^3.0.3
+  battery: ^2.0.1
+  flutter_spinkit: ^5.0.0
+  version: ^2.0.0
   direct_select_flutter: ^1.0.7
-  archive: ^2.0.13
-  code_editor: ^1.0.1
+  archive: ^3.1.2
+  code_editor:
+    git: https://github.com/hanerx/code_editor
   flutter_json_widget: ^1.0.2
-  pretty_json: ^1.1.0
-  lpinyin: ^1.1.1
-  gbk_codec: ^0.3.2
+  pretty_json: ^2.0.0
+  lpinyin: ^2.0.1
+  gbk2utf8: ^1.0.1
+  flutter_archive: ^3.0.0
+  sqlite3_flutter_libs: ^0.4.1
+  firebase_core: "^1.0.3"
+  firebase_analytics: "^8.0.0"
+  firebase_crashlytics: "^2.0.0"
+  firebase_auth: "^1.1.0"
   logger_flutter:
     git: https://github.com/ewertonrp/logger_flutter.git
   ipfs:
@@ -67,7 +73,7 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
