@@ -1,35 +1,50 @@
 import 'package:cached_network_image/cached_network_image.dart';
+import 'package:dcomic/model/comic_source/sourceProvider.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:dcomic/view/comic_detail_page.dart';
 import 'package:dcomic/view/subject_detail_page.dart';
+import 'package:provider/provider.dart';
 
-class CustomCard extends StatelessWidget {
+class HomePageCard extends StatelessWidget {
   final String imageUrl;
   final String title;
-  final String author;
+  final String subtitle;
   final int type;
   final String id;
 
-  CustomCard(this.imageUrl, this.title, this.author, this.type, this.id);
+  HomePageCard(this.imageUrl, this.title, this.subtitle, this.type, this.id);
 
   @override
   Widget build(BuildContext context) {
     // TODO: implement build
-    return Expanded(
-        child: FlatButton(
-      padding: EdgeInsets.all(0),
+    return TextButton(
+      style: ButtonStyle(padding: MaterialStateProperty.all(EdgeInsets.zero)),
       onPressed: () {
         switch (type) {
           case 1:
-            Navigator.push(context, MaterialPageRoute(builder: (context) {
-              return ComicDetailPage(id:id,title: title,);
-            }));
+            Navigator.push(
+                context,
+                MaterialPageRoute(
+                    builder: (context) {
+                      return ComicDetailPage(
+                        id: id,
+                        title: title,
+                        model: Provider.of<SourceProvider>(context)
+                            .activeSources
+                            .first,
+                      );
+                    },
+                    settings: RouteSettings(name: 'comic_detail_page')));
             break;
           case 5:
-            Navigator.push(context, MaterialPageRoute(builder: (context) {
-              return SubjectDetailPage(id);
-            }));
+            Navigator.push(
+                context,
+                MaterialPageRoute(
+                    builder: (context) {
+                      return SubjectDetailPage(id);
+                    },
+                    settings: RouteSettings(name: 'subject_detail_page')));
             break;
           default:
         }
@@ -65,7 +80,7 @@ class CustomCard extends StatelessWidget {
                   ],
                 ),
                 Text(
-                  '${author != null ? author : ''}',
+                  '${subtitle != null ? subtitle : ''}',
                   style: TextStyle(color: Colors.grey, fontSize: 14),
                   textDirection: TextDirection.ltr,
                   textAlign: TextAlign.start,
@@ -74,6 +89,6 @@ class CustomCard extends StatelessWidget {
               ],
             ),
           )),
-    ));
+    );
   }
 }
