@@ -17,47 +17,56 @@ class CategoryCard extends StatelessWidget {
     return FlatButton(
       padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
       child: Card(
-        elevation: 0,
+          elevation: 0,
           child: new Container(
-        child: Wrap(
-          children: <Widget>[
-           ClipRRect(
-             child:  AspectRatio(
-               child: CachedNetworkImage(
-                 imageUrl: '$cover',
-                 httpHeaders: {'referer': 'http://images.dmzj.com'},
-                 progressIndicatorBuilder: (context, url, downloadProgress) =>
-                     Center(child: CircularProgressIndicator(value: downloadProgress.progress),),
-                 errorWidget: (context, url, error) => Icon(Icons.error),
-               ),
-               aspectRatio: 1,
-             ),
-             borderRadius: BorderRadius.circular(5),
-           ),
-            Row(
-              children: [Expanded(
-                child: Text(
-                  '$title',
-                  textAlign: TextAlign.center,
-                  overflow: TextOverflow.ellipsis,
+            child: Wrap(
+              children: <Widget>[
+                ClipRRect(
+                  child: AspectRatio(
+                    child: CachedNetworkImage(
+                      imageUrl: '$cover',
+                      httpHeaders: {'referer': 'http://images.dmzj.com'},
+                      progressIndicatorBuilder:
+                          (context, url, downloadProgress) => Center(
+                        child: CircularProgressIndicator(
+                            value: downloadProgress.progress),
+                      ),
+                      errorWidget: (context, url, error) => Icon(Icons.error),
+                    ),
+                    aspectRatio: 1,
+                  ),
+                  borderRadius: BorderRadius.circular(5),
                 ),
-              )],
-            )
-          ],
-        ),
-      )),
+                Row(
+                  children: [
+                    Expanded(
+                      child: Text(
+                        '$title',
+                        textAlign: TextAlign.center,
+                        overflow: TextOverflow.ellipsis,
+                      ),
+                    )
+                  ],
+                )
+              ],
+            ),
+          )),
       onPressed: () {
         if (type == 0) {
-          Navigator.of(context).push(MaterialPageRoute(builder: (context) {
-            return CategoryDetailPage(tagId, title);
-          }));
+          Navigator.of(context).push(MaterialPageRoute(
+              builder: (context) {
+                return CategoryDetailPage(tagId, title);
+              },
+              settings: RouteSettings(name: 'category_detail_page')));
         } else {
-          Navigator.of(context).push(MaterialPageRoute(builder: (context) {
-            return NovelCategoryDetailPage(
-              categoryId: tagId,
-              title: title,
-            );
-          }));
+          Navigator.of(context).push(MaterialPageRoute(
+              builder: (context) {
+                return NovelCategoryDetailPage(
+                  categoryId: tagId,
+                  title: title,
+                );
+              },
+              settings: RouteSettings(name: 'novel_category_detail_page')));
         }
       },
     );
