@@ -1,4 +1,5 @@
 import 'package:dcomic/utils/log_output.dart';
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:logger/logger.dart';
 import 'package:sqflite/sqflite.dart';
 
@@ -46,7 +47,7 @@ class CookieDatabaseProvider {
   Future get<T>(String configName, {T defaultValue}) async {
     var batch = (await db).batch();
     batch.query("cookies", where: "key='$configName' AND provider='$name'");
-    var result = await batch.commit();
+    var result = await batch.commit()as List<dynamic>;
     try {
       switch (T) {
         case String:
@@ -60,9 +61,10 @@ class CookieDatabaseProvider {
         default:
           return result.first[0]['value'];
       }
-    } catch (e) {
+    } catch (e,s) {
+      FirebaseCrashlytics.instance.recordError(e, s, reason: 'cookieGetFailed:$configName, provider: $name');
       logger.w(
-          'action: configGetFailed, name: $name, configName: $configName, exception: $e');
+          'action: cookieGetFailed, name: $name, cookieName: $configName, exception: $e');
     }
     return defaultValue;
   }
