@@ -4,7 +4,6 @@ import 'package:dcomic/database/historyDatabaseProvider.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
 import 'package:dcomic/database/cookieDatabaseProvider.dart';
-import 'package:dcomic/database/database.dart';
 import 'package:dcomic/database/downloader.dart';
 import 'package:dcomic/http/UniversalRequestModel.dart';
 import 'package:dcomic/utils/soup.dart';
@@ -13,7 +12,6 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import 'package:dcomic/database/sourceDatabaseProvider.dart';
-import 'package:dcomic/http/http.dart';
 import 'package:dcomic/model/comic_source/baseSourceModel.dart';
 import 'package:dcomic/model/systemSettingModel.dart';
 import 'package:dcomic/utils/tool_methods.dart';
@@ -39,9 +37,8 @@ class DMZJSourceModel extends BaseSourceModel {
     if (comicId == null) {
       throw IDInvalidError();
     }
-    CustomHttp http = CustomHttp();
     try {
-      var response = await http.getComicDetail(comicId);
+      var response = await UniversalRequestModel.dmzjRequestHandler.getComicDetail(comicId);
       if (response.statusCode == 200) {
         var title = response.data['title'];
         var cover = response.data['cover'];
@@ -62,7 +59,7 @@ class DMZJSourceModel extends BaseSourceModel {
             .toList();
         var history = (await HistoryDatabaseProvider(this.type.name)
             .getReadHistory(comicId));
-        var lastChapterId = history == null ? null : history['lastChapterId'];
+        var lastChapterId = history == null ? null : history['last_chapter_id'];
         return DMZJComicDetail(
             comicId,
             description,
@@ -83,7 +80,7 @@ class DMZJSourceModel extends BaseSourceModel {
       if (_options.backupApi) {
         return this.getComicDetailBackup(comicId);
       } else {
-        throw ComicLoadingError();
+        throw ComicLoadingError(exception: e);
       }
     }
     return null;
@@ -91,8 +88,7 @@ class DMZJSourceModel extends BaseSourceModel {
 
   Future<ComicDetail> getComicDetailBackup(String comicId) async {
     try {
-      CustomHttp http = CustomHttp();
-      var response = await http.getComicDetailDark(comicId);
+      var response = await UniversalRequestModel.dmzjapiRequestHandler.getComicDetailWithBackupApi(comicId);
       if (response.statusCode == 200) {
         var data = jsonDecode(response.data)['data'];
         var title = data['info']['title'];
@@ -122,7 +118,7 @@ class DMZJSourceModel extends BaseSourceModel {
         ];
         var history = (await HistoryDatabaseProvider(this.type.name)
             .getReadHistory(comicId));
-        var lastChapterId = history == null ? null : history['lastChapterId'];
+        var lastChapterId = history == null ? null : history['last_chapter_id'];
         return DMZJComicDetail(comicId, description, 0, 0, title, cover, author,
             types, chapters, status, updateDate, options, lastChapterId, type);
       }
@@ -161,7 +157,7 @@ class DMZJSourceModel extends BaseSourceModel {
   @override
   // TODO: implement type
   SourceDetail get type => SourceDetail('dmzj', '默认-动漫之家', '默认数据提供商，不可关闭',
-      false, SourceType.LocalDecoderSource, false,true);
+      false, SourceType.LocalDecoderSource, false, true);
 
   @override
   Widget getSettingWidget(context) {
@@ -218,10 +214,10 @@ class DMZJSourceModel extends BaseSourceModel {
     // TODO: implement getFavoriteComics
     if (_userConfig.status == UserStatus.login) {
       try {
-        var response = await UniversalRequestModel()
-            .dmzjRequestHandler
+        var response = await UniversalRequestModel.dmzjRequestHandler
             .getSubscribe(
-                await SourceDatabaseProvider.getSourceOption<int>(type.name, 'uid'),
+                await SourceDatabaseProvider.getSourceOption<int>(
+                    type.name, 'uid'),
                 page);
         if (response.statusCode == 200) {
           var data = await HistoryDatabaseProvider('dmzj').getAllUnread();
@@ -274,7 +270,7 @@ class DMZJUserConfig extends UserConfig {
           await SourceDatabaseProvider.getSourceOption<String>('dmzj', 'uid');
       try {
         var response =
-            await UniversalRequestModel().dmzjRequestHandler.getUserInfo(_uid);
+            await UniversalRequestModel.dmzjRequestHandler.getUserInfo(_uid);
         if (response.statusCode == 200) {
           _avatar = response.data['cover'];
           _nickname = response.data['nickname'];
@@ -515,7 +511,7 @@ class DMZJUserConfig extends UserConfig {
                   _uid = list[0];
                   _status = UserStatus.login;
                   try {
-                    var response = await UniversalRequestModel()
+                    var response = await UniversalRequestModel
                         .dmzjRequestHandler
                         .getUserInfo(_uid);
                     if (response.statusCode == 200) {
@@ -549,8 +545,7 @@ class DMZJUserConfig extends UserConfig {
   Future<bool> login(String username, String password) async {
     // TODO: implement login
     try {
-      var response = await UniversalRequestModel()
-          .dmzjiRequestHandler
+      var response = await UniversalRequestModel.dmzjiRequestHandler
           .login(username, password);
       var responseData = response.data.toString();
       var data = jsonDecode(responseData.substring(1, responseData.length - 1));
@@ -568,8 +563,7 @@ class DMZJUserConfig extends UserConfig {
             _uid = list[0];
             _status = UserStatus.login;
             try {
-              var response = await UniversalRequestModel()
-                  .dmzjRequestHandler
+              var response = await UniversalRequestModel.dmzjRequestHandler
                   .getUserInfo(_uid);
               if (response.statusCode == 200) {
                 _avatar = response.data['cover'];
@@ -634,9 +628,8 @@ class DMZJWebSourceModel extends DMZJSourceModel {
     if (comicId == null) {
       throw IDInvalidError();
     }
-    CustomHttp http = CustomHttp();
     try {
-      var response = await http.getComicDetailWeb(comicId);
+      var response = await UniversalRequestModel.dmzjMobileRequestHandler.getComicDetailWeb(comicId);
       if (response.statusCode == 200) {
         var jsonString =
             RegExp('initIntroData(.*);').stringMatch(response.data);
@@ -682,7 +675,7 @@ class DMZJWebSourceModel extends DMZJSourceModel {
         }
         var history = (await HistoryDatabaseProvider(this.type.name)
             .getReadHistory(comicId));
-        var lastChapterId = history == null ? null : history['lastChapterId'];
+        var lastChapterId = history == null ? null : history['last_chapter_id'];
         return DMZJComicDetail(
             comicId,
             description,
@@ -711,8 +704,14 @@ class DMZJWebSourceModel extends DMZJSourceModel {
 
   @override
   // TODO: implement type
-  SourceDetail get type => SourceDetail('dmzj-web', '动漫之家网页',
-      '使用大妈之家移动网页版的接口，让漫画重新可以看了', true, SourceType.LocalDecoderSource, false,false);
+  SourceDetail get type => SourceDetail(
+      'dmzj-web',
+      '动漫之家网页',
+      '使用大妈之家移动网页版的接口，让漫画重新可以看了',
+      true,
+      SourceType.LocalDecoderSource,
+      false,
+      false);
 
   @override
   // TODO: implement options
@@ -891,8 +890,7 @@ class DMZJComicDetail extends ComicDetail {
 
   Future<void> getIfSubscribed() async {
     try {
-      var response = await UniversalRequestModel()
-          .dmzjRequestHandler
+      var response = await UniversalRequestModel.dmzjRequestHandler
           .getIfSubscribe(
               comicId,
               await SourceDatabaseProvider.getSourceOption(
@@ -980,9 +978,10 @@ class DMZJComicDetail extends ComicDetail {
   Future<void> updateUnreadState() async {
     // TODO: implement updateUnreadState
     bool login = await SourceDatabaseProvider.getSourceOption<bool>(
-        sourceDetail.name, 'login');
+        sourceDetail.name, 'login',
+        defaultValue: false);
     if (login) {
-      UniversalRequestModel().dmzjInterfaceRequestHandler.updateUnread(comicId);
+      UniversalRequestModel.dmzjInterfaceRequestHandler.updateUnread(comicId);
     }
     return super.updateUnreadState();
   }
@@ -996,13 +995,11 @@ class DMZJComicDetail extends ComicDetail {
     // TODO: implement isSubscribed
     if (isSubscribed) {
       SourceDatabaseProvider.getSourceOption(sourceDetail.name, 'uid').then(
-          (value) => UniversalRequestModel()
-              .dmzjRequestHandler
+          (value) => UniversalRequestModel.dmzjRequestHandler
               .addSubscribe(comicId, value));
     } else {
       SourceDatabaseProvider.getSourceOption(sourceDetail.name, 'uid').then(
-          (value) => UniversalRequestModel()
-              .dmzjRequestHandler
+          (value) => UniversalRequestModel.dmzjRequestHandler
               .cancelSubscribe(comicId, value));
     }
     _isSubscribed = isSubscribed;
@@ -1024,7 +1021,7 @@ class DMZJComic extends Comic {
   List<String> _pages = [];
   String _title;
   List<String> _chapterIdList;
-  PageType _type=PageType.url;
+  PageType _type = PageType.url;
   List _viewPoints = [];
 
   String _previous;
@@ -1074,9 +1071,8 @@ class DMZJComic extends Comic {
       await getComicWeb(comicId, chapterId);
       return;
     }
-    CustomHttp http = CustomHttp();
     try {
-      var response = await http.getComic(comicId, chapterId);
+      var response = await UniversalRequestModel.dmzjRequestHandler.getComic(comicId, chapterId);
       if (response.statusCode == 200) {
         _pageAt = chapterId;
         _pages =
@@ -1107,9 +1103,8 @@ class DMZJComic extends Comic {
   }
 
   Future<void> getComicWeb(String comicId, String chapterId) async {
-    CustomHttp http = CustomHttp();
     try {
-      var response = await http.getComicWeb(comicId, chapterId);
+      var response = await UniversalRequestModel.dmzjMobileRequestHandler.getComicWeb(comicId, chapterId);
       if (response.statusCode == 200) {
         var data = jsonDecode(response.data);
         _title = data['chapter_name'];
@@ -1140,8 +1135,7 @@ class DMZJComic extends Comic {
 
   Future<void> getComicBackup(String comicId, String chapterId) async {
     try {
-      CustomHttp http = CustomHttp();
-      var response = await http.getComicDetailDark(comicId);
+      var response = await UniversalRequestModel.dmzjapiRequestHandler.getComicDetailWithBackupApi(comicId);
       if (response.statusCode == 200) {
         var data = jsonDecode(response.data)['data'];
         var firstLetter = data['info']['first_letter'];
@@ -1155,7 +1149,7 @@ class DMZJComic extends Comic {
         while (true) {
           try {
             var item =
-                await http.getComicPage(firstLetter, comicId, chapterId, page);
+                await UniversalRequestModel.dmzjImageRequestHandler.getImage(firstLetter, comicId, chapterId, page);
             if (item.headers.value('Content-Type') == 'image/jpeg') {
               pages.add(
                   'http://imgsmall.dmzj.com/$firstLetter/$comicId/$chapterId/$page.jpg');
@@ -1253,17 +1247,17 @@ class DMZJComic extends Comic {
     if (login) {
       //获取UID
       var uid = await SourceDatabaseProvider.getSourceOption('dmzj', 'uid');
-      CustomHttp http = CustomHttp();
-      http.addHistoryNew(int.parse(comicId), uid, int.parse(chapterId),
+      UniversalRequestModel.dmzjInterfaceRequestHandler.addHistory(
+          int.parse(comicId), uid, int.parse(chapterId),
           page: page);
     }
     notifyListeners();
   }
 
   Future<void> getViewPoint() async {
-    CustomHttp http = CustomHttp();
     try {
-      var response = await http.getViewPoint(_comicId, _pageAt);
+      var response = await UniversalRequestModel.dmzjRequestHandler
+          .getViewpoint(_comicId, _pageAt);
       if (response.statusCode == 200) {
         print(
             "class: ComicModel, action: getViewPoint, responseCode: ${response.statusCode} chapterId:$_chapterId comicId:$_comicId responseData: ${response.data}");
