@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_easyrefresh/easy_refresh.dart';
 import 'package:dcomic/component/EmptyView.dart';
 import 'package:dcomic/component/LoadingCube.dart';
-import 'package:dcomic/component/comic/comicListTile.dart';
+import 'package:dcomic/component/comic/ComicListTile.dart';
 import 'package:dcomic/model/comicCategoryDetailModel.dart';
 import 'package:dcomic/view/comic_detail_page.dart';
 import 'package:provider/provider.dart';
@@ -76,7 +76,7 @@ class _CategoryDetailPage extends State<CategoryDetailPage> {
                                       builder: (context) => ComicDetailPage(
                                             id: item['id'].toString(),
                                             title: item['title'],
-                                          )));
+                                          ),settings: RouteSettings(name: 'comic_detail_page')));
                                 },
                                 headers: {'referer': 'https://m.dmzj.com'},
                               );
