@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_easyrefresh/easy_refresh.dart';
 import 'package:dcomic/component/EmptyView.dart';
 import 'package:dcomic/component/LoadingCube.dart';
-import 'package:dcomic/component/comic/comicListTile.dart';
+import 'package:dcomic/component/comic/ComicListTile.dart';
 import 'package:dcomic/http/http.dart';
 import 'package:dcomic/model/comicLatestUpdateModel.dart';
 import 'package:dcomic/view/ranking_page.dart';
@@ -71,7 +71,7 @@ class _LatestUpdatePage extends State<LatestUpdatePage> {
                                 builder: (context) => ComicDetailPage(
                                       id: item['id'].toString(),
                                       title: item['title'],
-                                    )));
+                                    ),settings: RouteSettings(name: 'comic_detail_page')));
                           },
                         );
                         // return list[index];
